Do not assume that bitcoin.conf file specifies rpcuser
and raise an exception if rpcpassword is not specified.

Comment: You can make RPC calls to bitcoind when bitcoin.conf specifies
just rpcpassword and leaves user name empty.  The class bitcoin.rpc.RawProxy
creates keys of the internal dictionary: conf = {}, based on the entries
in the configuration file.  Then  assumes that the rpcuser key exists and
uses it to create service_url.  This results in an error
when bitcoin.conf contains only rcppassword field and the following code
is executed:
>    proxy = bitcoin.rpc.RawProxy()