refactor chain param selection; fix naming for thread-local var
ranamed _frontend_class_meta -> _frontend_metaclass

This should have been two commits. This was basically work-in-progress
commit made 'in the flow', and this comment was added later.
It is unfortunate that this two changes are within one commit,
but now it is too much hassle to separate the code into separate commits.
The same might be true for a few other surrounding commits.
Sorry.