In mutable context, only substitute _from_instance, not all attributes
When the execution is withon the method of a mutable core class,
we only want all new instances to be mutable. This does not justify
substituting all the attributes of the immutable class for the
attributes of the mutable class, as the only static method that can
produce new instances is _from_instance, and other methods call this
method in the end.