Allow mutable/immutable variants in VectorSerializer
Add the bitcointx-specific bypass to VectorSerializer.stream_serialize()
that allows different classes in the serialized vector if the classes
only differ by their mutability, detected via _immutable_cls
bitcointx-specific attribute.

While this bypass is bitcointx-specific, this is OK because this is
only a relaxation of the check in a very specific circumstance,
and the meaning of the check is only 'extra precaution'.

Additional ergonomics justify this bypass, while supermajority of
erroneous cases are still restricted